fastapi-app/requirements.txt
fastapi
uvicorn
databricks-sdk
databricks-sql-connector
